+++ crates/maybe-uring/src/io/non_uring.rs
@@ -1,10 +1,10 @@
-use maybe_uring::{
+use crate::{
     buf::{IoBuf, IoBufMut},
+    io::{ReadOwned, WriteOwned},
     BufResult,
 };
-use tokio::io::{AsyncRead, AsyncWrite};
 
-use crate::{ReadOwned, WriteOwned};
+use tokio::io::{AsyncRead, AsyncWrite};
 
 impl<T> ReadOwned for T
 where
